# api/openapi.yaml
openapi: 3.1.0
info:
  title: Origami FarmOS API
  description: Offline-first, tablet-first farm operating system API for Origami Farms.
    See CONSTITUTION.md and product/MVP_SCOPE.md in the repository root for the product
    principles this API enforces (evidence-based recommendations, worker-observes/manager-decides
    separation, event-sourced history).
  version: 0.1.0
paths:
...
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      description: "Lets the mobile app restore a session from its stored token on\n\
        relaunch (\"log in once, stay logged in\") without re-sending a\npassword\
        \ \u2014 it just re-validates the bearer token and returns the\ncurrent profile."
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileOut'
      security:
      - HTTPBearer: []
...
  /api/v1/animals:
    get:
      tags:
      - animals
      summary: List Animals
      operationId: list_animals_api_v1_animals_get
      security:
      - HTTPBearer: []
      parameters:
      - name: farm_id
        in: query
        required: true
        schema:
          type: string
          title: Farm Id
      - name: species
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Species
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnimalOut'
                title: Response List Animals Api V1 Animals Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - animals
      summary: Create Animal
      description: "Registers a new animal \u2014 the start of its digital twin (tech\
        \ spec \xA713)."
      operationId: create_animal_api_v1_animals_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/animals/{animal_id}:
    put:
      tags:
      - animals
      summary: Update Animal
      description: "Full edit of an animal record (tech spec \xA712)."
      operationId: update_animal_api_v1_animals__animal_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: animal_id
        in: path
        required: true
        schema:
          type: string
          title: Animal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - animals
      summary: Move Animal
      operationId: move_animal_api_v1_animals__animal_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: animal_id
        in: path
        required: true
        schema:
          type: string
          title: Animal Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnimalMove'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - animals
      summary: Get Animal
      operationId: get_animal_api_v1_animals__animal_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: animal_id
        in: path
        required: true
        schema:
          type: string
          title: Animal Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnimalDigitalTwinOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnimalCreate:
      properties:
        tag:
          type: string
          title: Tag
        name:
          type: string
          title: Name
        species:
          type: string
          title: Species
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        sex:
          anyOf:
          - type: string
          - type: 'null'
          title: Sex
        birth_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birth Date
        acquisition_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acquisition Date
        acquisition_source:
          anyOf:
          - type: string
          - type: 'null'
          title: Acquisition Source
        sire_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Sire Tag
        dam_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Dam Tag
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Weight Kg
        color_markings:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Markings
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        status:
          type: string
          title: Status
          default: healthy
        health_score:
          type: integer
          title: Health Score
          default: 100
        pregnant:
          type: boolean
          title: Pregnant
          default: false
        pregnancy_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pregnancy Days
        lactating:
          type: boolean
          title: Lactating
          default: false
        lactation_cycle:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lactation Cycle
        purchase_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Purchase Cost
        current_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Value
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - tag
      - name
      - species
      title: AnimalCreate
      description: "The full Add-Animal record (tech spec \xA713). Only tag/name/species\n\
        are required \u2014 a farmer standing in a barn should be able to register\n\
        an animal in three fields and fill in the rest later."
...
    AnimalMove:
      properties:
        location_label:
          type: string
          title: Location Label
      type: object
      required:
      - location_label
      title: AnimalMove
    AnimalOut:
      properties:
        id:
          type: string
          title: Id
        farm_id:
          type: string
          title: Farm Id
        tag:
          type: string
          title: Tag
        name:
          type: string
          title: Name
        species:
          type: string
          title: Species
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        sex:
          anyOf:
          - type: string
          - type: 'null'
          title: Sex
        birth_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birth Date
        status:
          type: string
          title: Status
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        health_score:
          type: integer
          title: Health Score
        pregnant:
          type: boolean
          title: Pregnant
        pregnancy_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pregnancy Days
        lactating:
          type: boolean
          title: Lactating
        lactation_cycle:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lactation Cycle
        withdrawal_until:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Withdrawal Until
        withdrawal_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Withdrawal Reason
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Weight Kg
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        acquisition_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acquisition Date
        acquisition_source:
          anyOf:
          - type: string
          - type: 'null'
          title: Acquisition Source
        sire_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Sire Tag
        dam_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Dam Tag
        color_markings:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Markings
        purchase_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Purchase Cost
        current_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Value
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        active:
          type: boolean
          title: Active
          default: true
      type: object
      required:
      - id
      - farm_id
      - tag
      - name
      - species
      - status
      - health_score
      - pregnant
      - lactating
      title: AnimalOut
    AnimalUpdate:
      properties:
        tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        species:
          anyOf:
          - type: string
          - type: 'null'
          title: Species
        breed:
          anyOf:
          - type: string
          - type: 'null'
          title: Breed
        sex:
          anyOf:
          - type: string
          - type: 'null'
          title: Sex
        birth_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Birth Date
        acquisition_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Acquisition Date
        acquisition_source:
          anyOf:
          - type: string
          - type: 'null'
          title: Acquisition Source
        sire_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Sire Tag
        dam_tag:
          anyOf:
          - type: string
          - type: 'null'
          title: Dam Tag
        location_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Label
        group_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Group Name
        weight_kg:
          anyOf:
          - type: number
          - type: 'null'
          title: Weight Kg
        color_markings:
          anyOf:
          - type: string
          - type: 'null'
          title: Color Markings
        photo_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Photo Path
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
        health_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Health Score
        pregnant:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Pregnant
        pregnancy_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Pregnancy Days
        lactating:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Lactating
        lactation_cycle:
          anyOf:
          - type: integer
          - type: 'null'
          title: Lactation Cycle
        purchase_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Purchase Cost
        current_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Current Value
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
        active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Active
      type: object
      title: AnimalUpdate
      description: "Full edit (tech spec \xA712). `location_label` is here too, so\
        \ the\ndedicated Move action and a general edit share one code path."
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
        device_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Id
      type: object
      required:
      - email
      - password
      title: LoginRequest
    LoginResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        expires_in_minutes:
          type: integer
          title: Expires In Minutes
        user:
          $ref: '#/components/schemas/UserProfileOut'
      type: object
      required:
      - access_token
      - expires_in_minutes
      - user
      title: LoginResponse
...
    UserProfileOut:
      properties:
        id:
          type: string
          title: Id
        farm_id:
          type: string
          title: Farm Id
        name:
          type: string
          title: Name
        email:
          anyOf:
          - type: string
          - type: 'null'
          title: Email
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        role:
          type: string
          title: Role
        department:
          anyOf:
          - type: string
          - type: 'null'
          title: Department
        language:
          type: string
          title: Language
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - farm_id
      - name
      - role
      - language
      - active
      title: UserProfileOut
      description: "Canonical user-facing shape \u2014 used by /auth/login, /auth/me,\
        \ and\nGET /users, so the mobile app has exactly one profile shape to parse\n\
        regardless of which endpoint produced it.\n\n`department` (nullable) is a\
        \ mobile-UI/task-assignment concern only \u2014\nit drives which nav entries\
        \ an employee sees and who a task can be\nassigned to for a given area; it\
        \ never gates a backend permission\n(that's still `role`, checked by api/deps.py)."
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer